+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add support for client hints. ([#1752](https://github.com/getsentry/relay/pull/1752))
+- Add count transactions toward root project. ([#1734](https://github.com/getsentry/relay/pull/1734))
 
 **Bug Fixes**: